backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector DB
qdrant-client>=1.7.0
chromadb>=0.4.0

# LLM & Embedding
openai>=1.3.0
litellm>=1.0.0

# Document processing
python-docx>=1.0.0
python-pptx>=0.6.0
pdfplumber>=0.10.0

# HTTP
httpx>=0.25.0
requests>=2.31.0

# Auth
PyJWT>=2.8.0

# Utils
python-multipart>=0.0.6
